Got tired of trying to make things nicely, went with brute force.
I had something going with my spreadsheets, starting to see a pattern, but couldn't find the logical step to go from 3 to 4, so couldn't take it from there. Decided that brute force and getting ALL the solutions (including reflections and rotations because I'm too lazy to do the math). Yes there are efficiencies that could be added, for example not having points double check themselves, but again, this was frustration coding.